ClientNotificationNodeQueue maximum queue size
Change the default value for the queue size and number of notifications
that are processed at once. Several projects are already using this
larger value, the old value is no longer up to date. In addition, change
the log level from Warn to Error so that no project misses this error.

368114