FocusManager: fix wrong focus after dragging an element
The global mouse down listener of the focus manager is responsible to
allow or prevent focus gain on mouse down. If an element is draggable,
the focus gain cannot be prevented without preventing the dragstart
event, too.

Therefore, the focus manager allows the focus gain but has to restore
the focus afterward, so that the previously focused element will be
focused again. This functionality broke with commit
f318d652e5b6ec5b028f1f08fe762ead278e3951 (#381281).

To fix it, the logic from FocusContext to prevent the focus gain on
elements with tabindex=-2 has to be integrated into the global
mouse down handler. One advantage of this approach is that such
elements now don't get the focus at all, not even temporary, except
if a draggable element is clicked. As explained, the focus gain must
be allowed and may temporarily be set on an element with tabindex=-2.

Also:
- removed unnecessary HTMLElement generic of the JQuery type because
  it is the default value
- simplified checks in isFocusableByMouse and isDraggable because
  closest also considers the current element not only the parents.

397963