UiNotifications: don't poll during tests
If a component subscribes for a ui notification topic, the ui
notification polling starts. These poll requests will fail and new poll
requests will be scheduled, which leads to a lot of noise.

During tests, ui notifications are now disabled by default. If a
specific test needs it to be enabled, the mock can be removed by
calling UiNotificationMock.unregister(). To not influence other specs,
it should be registered again afterward.

Also:
- uiNotifications is now an instance of the class UiNotifications,
  so it can be replaced using the object factory.
- UiNotificationSystem can now be replaced using the object factory.

388703